### 2025-12-15 #191 @luraess
Excellent! Thanks a lot for being proactive @kellertuer ! Would @Nimrais (Mykola) and @Affie (Johannes) be willing to review this contribution? Thanks all for your feedback.
