### 2025-12-16 #191 @Nimrais
@luraess Yeah, I would like to review the paper and the package.
