### 2026-01-07 #191 @kellertuer
We could maybe also ask from the list of JuliaCon proceedings reviewers above:

* @vbrachetta
* @jbytecode
* @mzagorowska
* @nrummel
* @AFeuerpfeil

Though differential geometry / Lie groups is not your main focus, numerics / optimization / differential equations are areas this package should/could be used in, so they do fit the audience this proceedings paper addresses.
