### 2026-01-07 #191 @AFeuerpfeil
@kellertuer I would be open to reviewing it, as I have had some courses on differential geometry and Lie groups.
