### 2026-01-08 #191 @kellertuer
@jbytecode no worries, that is why I already wrote the request carefully. I am aware this area is a bit niche.
@AFeuerpfeil great. thanks!
