### 2026-01-08 #195 @editorialbot
[REVIEW]: Groups and smooth geometry using LieGroups.jl
**Submitting author:** @kellertuer (<a href="http://orcid.org/0000-0001-8342-7218">Ronny Bergmann</a>)
**Repository:** https://github.com/JuliaManifolds/LieGroups.jl
**Branch with paper.md** (empty if default branch): JuliaCon2025Procedings
**Version:** v0.1.9
