### 2026-01-08 #195 @luraess
Dear all, thanks for your help with the revision of this JCon Proceedings submission. As suggested by the Editorial Bot, please generate your reviewer checklist using

Ideally, you can swiftly address all the points marking them as checked when completed so we can all track progress. Good practice is to open directly issues or PRs in the related code repository and link them to a succinct review as comment in this conversation.

Note that there is no need for an explicit "statement of need", but the content of such statement should be somehow included in the first part of the manuscript.

Thanks!
