### 2026-01-08 JuliaManifolds/LieGroups.jl#87 @AFeuerpfeil
Mathematical definition of tangent space and smooth manifold
1. The tangent space is defined using “derivatives of smooth curves,” but “derivative of a curve on a manifold” is not defined in the text, which makes the definition feel circular. To make the paper self-contained and to clarify what structures the implementation assumes, define the derivative/velocity of a smooth curve via a chart, and then define tangent vectors as equivalence classes of curves with the same chart-derivative at the point.
2. The definition of a smooth manifold omits standard topological assumptions (Hausdorff and second-countable). Without stating these explicitly, the definition unintentionally includes pathological “locally Euclidean” spaces that are very likely outside the intended scope and not something the package can reasonably support, so adding these conditions would better delimit what users can expect to be implementable.
