### 2026-01-08 #195 @AFeuerpfeil
I suggest publication after the minor revision of this issue related to the mathematical definitions in the paper: https://github.com/JuliaManifolds/LieGroups.jl/issues/87
