### 2026-01-08 JuliaManifolds/LieGroups.jl#87 @kellertuer
Thanks for the super fast work on a review of the paper!

Both are valid points and we tried to find a good balance between a super rigorous mathematical paper (but meybe technical and lengthy) versus an easy-to-reed but maybe every now and then not 100% rigorous one.

Concerning 1, sure that is very intuitive. To make that rigorous, yes we would need to define charts and then take the derivative of the composition of the curve and the chart. Until now we do not yet define charts (that is exactly what we wanted to avoid for technical reasons). Doing that thoroughly would take half a page (or one column) I would assume.
Shall we do all that definition of charts and transition maps or is it ok to write that informally?
(Maybe along the lines of: locally a manifold is isomorphic to R^n, like the sphere “locally looks like” a plane, take that as a chart,...?)

Concerning 2, sure, we can mention both points – shall we also define both terms or is it ok to refer to corresponding literature doing that?
