### 2026-01-08 JuliaManifolds/LieGroups.jl#87 @AFeuerpfeil
Thanks for the quick response! I completely agree that there is a fine line to walk here between a maths and a software paper.

However, given the prominence of the JuliaManifolds ecosystem, I feel that the definitions for fundamental objects like Lie groups and Lie algebras need to be mathematically sound. Precise definitions serve as a reliable reference for the community, whereas imprecise ones can cause confusion about the software's intended scope. I am perfectly fine with skipping proofs or deep technicalities (like the existence proof of the exponential map), but the what needs to be correct.

To both points:
I don't think you need to define all terms you are using (like chart, Hausdorff, topological space, etc), as the reader either knows these or can look them up. My main point is to have correct definitions. Therefore:
1. I would be fine with defining the atlas (and charts) informally or just referencing a source, but then give a proper definition of the tangent space.
2. I would not define them, but add them to the definition of a manifold being a topological space that is Hausdorff, second-countable and there exists a set of charts to an R^n.
