### 2026-01-08 JuliaManifolds/LieGroups.jl#87 @kellertuer
I totally agree on avoiding too vague definitions that rely too much on intuition.

Besides charts I would also want to mention the change of charts and that that makes the derivative of the curve really precise, since it is then independent of charts. I also think at the change of charts we decided to take take the current short route we went for ;)
But I will take a while to think about a reasonable level of detail of all this and propose a PR for that (and mention you there) soon.

For the second point – yes that sounds reasonable, will do that.
