### 2026-01-09 cce7c91
Bumpo version in JuliaCon2025 paper
