+++ paper/paper.tex
@@ -14,7 +14,7 @@
 \newcommand{\term}[1]{\emph{#1}}
 \newcommand{\e}{\ensuremath{\mathrm{e}}}
 \newcommand{\im}{\ensuremath{\mathrm{i}}}
-\newcommand{\LieGroupsVersion}{0.1.8}
+\newcommand{\LieGroupsVersion}{0.1.9}
 %
 % Define muted colours.
 \definecolor{TolMutedBlue}{HTML}{332288}
