### 2026-01-09 e207ae0
Add a reference, the book by Boumal.
