+++ paper/ref.bib
@@ -23,6 +23,18 @@
     Year      = {2022},
 }
 
+@book{Boumal:2023,
+    Title     = {An Introduction to Optimization on Smooth Manifolds},
+    Author    = {Boumal, Nicolas},
+    Year      = {2023},
+    Month     = mar,
+    Edition   = {First},
+    Publisher = {Cambridge University Press},
+    Doi       = {10.1017/9781009166164},
+    ISBN      = {978-1-00-916616-4},
+    URL       = {https://www.nicolasboumal.net/#book}
+}
+
 @article{bezanson2017julia,
   title={Julia: A fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
