### 2026-01-09 f412f1d
Add the notions of chart and atlas
