### 2026-01-09 e5d0593
Add a more precise definition of tangent spaces, especially regarding the derivative of curves.
