### 2026-01-09 JuliaManifolds/LieGroups.jl#87 @kellertuer
Here are the changes I made:
* I merged `LieGroups.jl#main` and bumped the version number in the manuscript to 0.1.9
* I added a reference to provide more details on manifolds in https://github.com/JuliaManifolds/LieGroups.jl/commit/e207ae0549f52e42ff3d589f316d518779ec3161
* I introduced charts, the transition maps, the notion of an atlas and when it is smooth (and hence we have a smooth manifold) in https://github.com/JuliaManifolds/LieGroups.jl/commit/f412f1de586e1cdd8872bd55f42b080f49fe56e7
* I mentioned (including a reference) second-countable and Hausdorff in that commit as well
* I added a formal definition of the velocities / a derivative of a curve in https://github.com/JuliaManifolds/LieGroups.jl/commit/e5d05931f8a79ef132bf4a93af3979ba1cab1255 and added an example at the end continuing the 2-sphere example from before.

The final version even still fits 8 pages, see https://github.com/JuliaManifolds/LieGroups.jl/blob/94ce5f97a2db16732bdc7a222cb711a18022826b/paper/paper.pdf

Oh and I could have done this as a PR to the Proceedings branch, sorry. But I hope the changelog here is also good enough to follow what was changed.
In the end I might change the “column break” on the last page a bit (after the other review) so that the enumeration is not split between both columns as it is now.

Let me know what you think!
