### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
Peer Review: Groups and Smooth Geometry using LieGroups.jl
## Overall Assessment

I begin by complimenting the authors on developing such a valuable package that addresses a significant need in the Julia ecosystem. Many colleagues in my laboratory work on applications that would benefit substantially from `LieGroups.jl`. The contribution to the Julia ecosystem is monumental, and I thank the authors for their effort.

**I do not have any immediate blocking recommendations regarding the code itself.**

Regarding the repository in general, a clearly visible guideline explaining how to contribute to the package should be added either in the documentation or in the README file.

However, I have several suggestions for improving the paper's presentation and completeness.

---

## Comments

### 1. Target Audience and Mathematical Background

#### Comment

The current presentation of the mathematical background (Section 2) presents a challenge in terms of target audience clarity. The exposition jumps in level of abstraction, which, while acceptable for a short-format paper, creates ambiguity about the intended readership:

1. **If targeting engineers** who will use Lie groups in practice: The mathematical background in its current form may read as abstract noise without clear connections to practical applications.

2. **If targeting mathematicians** who wish to see how these abstract tools translate to practice: They likely already possess this background knowledge and do not require it in this form.

#### Recommendation

I suggest restructuring the mathematical background to align more directly with the first application example (the Bézier curves on SE(2)). The Bézier curve example is quite visual and intuitive—it effectively demonstrates the package's capabilities to someone familiar with manifolds and/or Bézier curves. By mapping the mathematical concepts directly onto this concrete example, engineers would gain better intuition, while mathematicians would see the practical relevance immediately.

Additionally, consider citing introductory Lie group material for readers who need deeper background, rather than attempting to cover it within the paper itself. While citation [8] (Hilgert & Neeb's Structure and Geometry of Lie Groups) is substantial and certainly sufficient for understanding the package, it may be too comprehensive for engineers seeking a quick entry point. A more accessible reference such as:

> Solà, J., Deray, J., & Atchuthan, D. (2018). A micro Lie theory for state estimation in robotics. arXiv preprint arXiv:1812.01537.

would serve engineering practitioners better, as it provides a concise, application-oriented introduction to Lie theory specifically tailored for robotics and state estimation contexts.

---

### 2. Performance Comparison with Existing Tools

Since this paper is a full proceedings paper rather than an extended abstract, I would strongly encourage the inclusion of a performance comparison with existing tools in the robotics community. Specifically, the Python package **navlie** (Cossette et al., 2023) is widely used for state estimation on Lie groups.

Both `LieGroups.jl` and `navlie` can be used to implement Extended Kalman Filters on manifolds as demonstrated in the authors' related work on geometric Kalman filtering (arXiv:2506.01086). A benchmark comparing:

1. Computational performance (execution time, memory usage)
2. Ease of implementation for equivalent filtering tasks
3. Numerical accuracy on standard test problems

would significantly strengthen the paper's contribution by demonstrating the practical advantages of adopting a Julia-based solution over a Python alternative.

Even if `LieGroups.jl` were to perform worse than `navlie` in its current form, I do not believe this would undermine the package in any sense—`LieGroups.jl` is a much more general solution. However, such a comparison would provide robotics practitioners considering adoption with an immediate answer as to whether they should invest time in exploring the package.

---

## Recommendation

I recommend **minor revision** of the manuscript. The package itself represents excellent work, and the paper would benefit from:

1. Reorganizing the mathematical background to connect more directly with the application examples
2. Including performance benchmarks against comparable tools (e.g., `navlie`)

These revisions would make the paper more accessible to its intended audience and provide stronger empirical evidence for the package's practical value.

---

## References Consulted

- Baran, M. & Bergmann, R. (2025). *A geometric perspective of state estimation using Kalman filters.* arXiv:2506.01086

- Cossette, C. C., Cohen, M., Korotkine, V., Bernal, A. D. C., Shalaby, M. A., & Forbes, J. R. (2023). *navlie: A Python package for state estimation on Lie groups.* In 2023 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS), pp. 5282–5287. IEEE.

- Solà, J., Deray, J., & Atchuthan, D. (2018). *A micro Lie theory for state estimation in robotics.* arXiv:1812.01537
