### 2026-01-09 #195 @Nimrais
I suggest publication after the minor revision. You can read my review in the issue here https://github.com/JuliaManifolds/LieGroups.jl/issues/88.
