### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
P.S. I would like to clarify that my comments are suggestions, not hard blockers. They do not need to be implemented in any particular way. My recommendations are simply examples of how they could be addressed. I understand that in the mathematical community not everyone agrees with the Micro Lie Theory exposition due to a lack of rigour in some places; however, I personally find a 700-page book as a citation a bit excessive. I am certainly not hard blocking the authors in any way on this matter. Any concise paper with a comprehensive introduction would suffice; this paper is simply one I like to recommend as a starting point for engineers.
