### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @kellertuer
Hi, thanks for the detailed review and feedback and your very kind words, as well as for doing the review super-fast. That is much appreciated.

Let me start with the last comment: We do use the book by Hilgert and Neeb as a main reference throughout the whole documentation. While that book is a bit long (and maybe technical to the non-so-abstract people), in every single documentation string we narrow it down as best as we can. For example in https://juliamanifolds.github.io/LieGroups.jl/stable/interface/group/#Base.adjoint-Tuple{AbstractLieGroup,%20Any,%20Any} we refer to a section that is two pages.
While the overall book is a bit long, these single blocks contain then exactly what we need.

Yes, I personally consider the micro Lie Theory paper (a) too short and (b) too less precise as a reference. One thing that could be done is, to relate the current definitions to elements int that paper. My problem with that is, that for me the paper is so less precise, that I struggle every now and then to pin point these relations.

For the remaining remarks let me first try to narrow down what best to do before sitting down and rewriting the paper. I will try my best to take the comments here into account after clarification (see also https://github.com/JuliaManifolds/LieGroups.jl/issues/87 where we did the same already :))

> Regarding the repository in general, a clearly visible guideline explaining how to contribute to the package should be added either in the documentation or in the README file.

The https://github.com/JuliaManifolds/LieGroups.jl/blob/main/CONTRIBUTING.md file is included in the documentation at https://juliamanifolds.github.io/LieGroups.jl/stable/contributing/ – is that enough or should this link _also_ be added to the Readme?

> The current presentation of the mathematical background (Section 2) presents a challenge in terms of target audience clarity. The exposition jumps in level of abstraction, which, while acceptable for a short-format paper, creates ambiguity about the intended readership:
>
> 1. **If targeting engineers** who will use Lie groups in practice: The mathematical background in its current form may read as abstract noise without clear connections to practical applications.
> 2. **If targeting mathematicians** who wish to see how these abstract tools translate to practice: They likely already possess this background knowledge and do not require it in this form.

I am unsure how to improve this, since following #87 we even _increased_ the mathematical exactness. I am also a bit lost on what “practical applications” refers to here? We use the sphere as a guiding example for manifolds and SO(n), n=2,3, as a guiding example for a Lie group. So those _are_ the practical examples you refer to we should Taylor the text to?

On the other hand the whole package follows this level of abstraction as well, so the section 2 is roughly the level the documentation uses for now as well.

As already mentioned, I do not want to cite Micro Lie theory. I find that confusing. They try to define a plus and a minus that are not well-defined and their left and right Jacobian are mathematically also not correct.

> Since this paper is a full proceedings paper rather than an extended abstract, I would strongly encourage the inclusion of a performance comparison with existing tools in the robotics community. Specifically, the Python package **navlie** (Cossette et al., 2023) is widely used for state estimation on Lie groups.

We have about 6-8 lines of text left in the 8 page limit.
Usually when I try to use Python I get haunted by package dependency hell to an extend that I fear if I would try to write any benchmarking against navlie, this will postpone the proceedings paper by several months. I also have never used nave before.

If someone wants to do such a benchmarking I am open to helping from the Julia side, but if this proceedings paper should really include Python code (as a JuliaCon proceedings paper) I would be very unhappy and maybe even refrain from continuing here.

So both for the paper length and me really disliking Python I would kindly ask to not having to do this. This proceedings paper is really containing mainly the content of the talk at JuliaCon this year (plus 2 short code examples we came up with in November).

Sorry for a bit of negative comments on your feedback, I hope we still find a way to make this short paper possible.
