### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @kellertuer
PS. A short glance at https://github.com/decargroup/navlie seems to indicate it is better to be compare with https://github.com/JuliaManifolds/GeometricKalman.jl or KalmanFilter.jl, since LieGroups.jl itself does to zero extend to state estimation. It is really (just) to define Lie groups.

_edit:_ If I understand the section at https://decargroup.github.io/navlie/_build/html/tutorial/lie_groups.html#Define-the-State correctly, one has to implement the Lie group at hand oneselves before using Navlie? What they call plus/minus would be exp/log we provide in the package, but if I want to compare to them, _they_ should provide that.
On the other hand they have filters – LieGroups.jl does not provide those and I would have to program them in Jlia myself then? I am unsure what to compare even _if_ I would get navlie to work locally (which I do not expect to get working).
