### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @mateuszbaran
Thank you for a thorough evaluation.

We've analyzed the micro Lie theory paper and I agree with Ronny's evaluation that this paper is not good. It is inconsistent, skips too much theory, focuses on some wrong abstractions and generally does not offer the right impression about state estimation. Our preprint (arXiv:2506.01086) was largely written to fill some of these gaps and make some corrections. It is very unfortunate that there is currently no good introductory text about non-Euclidean state estimation.

As state estimation is beyond the scope of this package, I think it would be more appropriate to compare LieGroups.jl with https://github.com/decargroup/pymlg which is used by navlie. I think I could do that.
