### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> As state estimation is beyond the scope of this package, I think it would be more appropriate to compare LieGroups.jl with https://github.com/decargroup/pymlg which is used by navlie. I think I could do that.

Yeah comparison with `pymlg` would much better, I did not know about this package to be honest :). But as turned out it's coming from the same group and I did use it from there.
