### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> The https://github.com/JuliaManifolds/LieGroups.jl/blob/main/CONTRIBUTING.md file is included in the documentation at https://juliamanifolds.github.io/LieGroups.jl/stable/contributing/ – is that enough or should this link also be added to the Readme?

Yes, a one liner in the README would be nice.
