### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> I am unsure how to improve this, since following https://github.com/JuliaManifolds/LieGroups.jl/issues/87 we even increased the mathematical exactness. I am also a bit lost on what “practical applications” refers to here? We use the sphere as a guiding example for manifolds and SO(n), n=2,3, as a guiding example for a Lie group. So those are the practical examples you refer to we should Taylor the text to?

> We've analyzed the micro Lie theory paper and I agree with Ronny's evaluation that this paper is not good. It is inconsistent, skips too much theory, focuses on some wrong abstractions and generally does not offer the right impression about state estimation. Our preprint (arXiv:2506.01086) was largely written to fill some of these gaps and make some corrections. It is very unfortunate that there is currently no good introductory text about non-Euclidean state estimation.

I checked the other review, and it would be impossible to address both reviews together considering the page limit. I will leave this decision to you.

My belief is that in a software package paper, where we are not proving theorems, some technicalities could slide in order to address a broader community. Even if some foundations are imprecise at the start, such as in the Micro Lie Theory paper, it is very hard to convince an engineer to use a package when you begin your background section with a reference to a large book. In my experience, most of them would simply close the paper at that point. That said, this could be perfectly fine if you are addressing a different audience.
