### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @kellertuer
> Even if some foundations are imprecise at the start, [...]

I do not agree here. Keeping things a bit short is ok in order to have a focussed introduction. But it should be precise. And especially never be wrong.

> it is very hard to convince an engineer to use a package when you begin your background section with a reference to a large book

From that book we in practice use one chapter (cf. https://github.com/search?q=repo%3AJuliaManifolds%2FLieGroups.jl%20HilgertNeeb&type=code) I thought we might have used Ch. 13 as well, so maybe 2; but it is the one that is focussed and precise.

Since I come from the math side (-ish), that is what I can offer.

If an engineer wants to step in and write a tutorial “How LieGroups.jl relates to Micro-Lie” I am willing to try to help pointing out imprecisions in these relations/connections and match terms/names, but I would not attempt such a tutorial on my own; as pointed out before, I think that paper is not good and builds a lot of wrong intuitions.

> I checked the other review, and it would be impossible to address both reviews together considering the page limit. I will leave this decision to you.

I do agree (for you maybe sadly) very much with the other review and the remark was something I happily clarified and hope it is to the other reviewer liking. Hence I would prefer to stick to that choice.
