### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> I do agree (for you maybe sadly) very much with the other review and the remark was something I happily clarified and hope it is to the other reviewer liking. Hence I would prefer to stick to that choice.

I am not sad at all why would I be :).

> If an engineer wants to step in and write a tutorial “How LieGroups.jl relates to Micro-Lie” I am willing to try to help pointing out imprecisions in these relations/connections and match terms/names, but I would not attempt such a tutorial on my own; as pointed out before, I think that paper is not good and builds a lot of wrong intuitions.

Indeed, if someone were to write a precise yet concise introduction, that would be very nice. I think we can close the discussion on the first part of my review regarding the Micro Lie Theory citation. I will not fight for it in any way, as it is not a perfect paper by any means. If you don't have any small paper in mind to replace it with—perhaps because one simply does not exist—then I can consider this point of my review addressed.
