### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @kellertuer
> Yes, a one liner in the README would be nice.

I did add a line on the main branch and for completeness also merged that to the JuliaConProc branch.
