### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> From that book we in practice use one chapter (cf. https://github.com/search?q=repo%3AJuliaManifolds%2FLieGroups.jl%20HilgertNeeb&type=code) I thought we might have used Ch. 13 as well, so maybe 2; but it is the one that is focussed and precise.

Yeah to be honest, if you can precise your citation in the paper to a specific chapter of the book that would be much less intimidating already.
