### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> So both for the paper length and me really disliking Python I would kindly ask to not having to do this. This proceedings paper is really containing mainly the content of the talk at JuliaCon this year (plus 2 short code examples we came up with in November).

Returning to the second part of my review. My point of view is the following: we have "Results: are the results presented and compared to approaches with similar goals?" in the checklist, so if we can compare, we probably should, right?

> Usually when I try to use Python I get haunted by package dependency hell to an extend that I fear if I would try to write any benchmarking against navlie, this will postpone the proceedings paper by several months.

I did not have any problems installing the package, and I also verified that their example runs successfully though I had some issues getting their plotting functions to work with their latest commit: https://github.com/Nimrais/ReproduceNavlie.

> As state estimation is beyond the scope of this package, I think it would be more appropriate to compare LieGroups.jl with https://github.com/decargroup/pymlg which is used by navlie. I think I could do that.

I agree with @mateuszbaran that it would be more appropriate to compare with `pymlg`; however, this package is not registered. I have also never used it as a standalone tool—I tried it for the first time today. So if you have even a minimal idea of how to compare with it, that would be nice.
