### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @kellertuer
> [...] so if we can compare, we probably should, right?

My personal view point is: I personally can not. But Mateusz might give it a try.
I still feel that comparing to exactly one package, that by now was not even on your radar is more like a bit of an ad-hoc comparison. What about C libraries? Fortran? Rust? Do they exist? Can we compare?
We for example did a very thorough literature/software review and benchmark comparison in the paper about Manifolds.jl – but that was both a much much longer undertaking as well as a full Journal paper. To me personally this feels out of scope for a package that is as new as LieGroups.jl

> Yeah to be honest, if you can precise your citation in the paper to a specific chapter of the book that would be much less intimidating already.

Currently the paper has 4 placed where we cite Hilgert-Neeb.
* In the introduction at https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L59-L60 where besides that book we cite the two books by Chirikjian also without more detail (in sum 800+ pages) since this is meant as an overview
* In the very first lines of the math section at https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L88 together with one reference for manifolds – that is both meant as the generic reference.
* for the concrete cases of the Lie group definition (https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L147 with exact reference to the definition and in the overview for Lie algebra referring to their chapter 5 https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L311 which – yes – is indeed 40 pages discussing that.

The first is meant as an overview and Chirikjian is on purpose a far more engineering tailored reference as a “counter-weight” – and then even a longer reference.
Also – there is no free lunch. Working with Lie groups and Lie algebras does require a bit of diving into the corresponding theory, like it is with many topics.
For notation in the second bullet point, I am not sure how to rephrase that, we use the notation of that book. That is what it is; at least the first half of their index can serve as a good notation overview as well.;
References 3 and 4 are as precise as I can get them.
