### 2026-01-09 JuliaManifolds/LieGroups.jl#88 @Nimrais
> We have about 6-8 lines of text left in the 8 page limit.

Could you please point me to where you found that the page limit is 8 pages? I checked the author's guide, and it clearly states that papers can be up to 10 pages https://juliacon.github.io/proceedings-guide/author/. It is possible that there are two different places where JuliaCon provides different information.

> My personal view point is: I personally can not. But Mateusz might give it a try.

I understand that this comparison is ad-hoc, and I agree it would be unreasonable to ask for a table as thorough as Table 1 or a comparison as precise as Figure 3 in your Manifolds.jl paper. However, the proceedings guide does state that full papers should compare the work _to other approaches in the field:_

"Compared to an extended abstract, a full paper presents more of the background and context motivating the work. It compares the work to other approaches taken in the field and gives some additional insights on the conference contribution. Use cases back up the work by showing how it can be used."

If providing such a comparison feels out of scope for a package as new as LieGroups.jl, perhaps an extended abstract would be a more suitable format?

I do not mean this as a criticism—I simply want to ensure that the submission format aligns with its content and your available resources.
