### 2026-01-10 JuliaManifolds/LieGroups.jl#88 @kellertuer
> Could you please point me to where you found that the page limit is 8 pages?

Ehm, in my notes ;) and those were probably wrong then, because you are right it is 5-10; sorry for that.

Concerning

> "Compared to an extended abstract, a full paper presents more of the background and context motivating the work. It compares the work to other approaches taken in the field and gives some additional insights on the conference contribution. Use cases back up the work by showing how it can be used."

I feel a _benchmark_ is out of scope, because LieGroups.jl is very new and hence not yet tailored that much towards performance. I also do understand this clarification for the full paper as possibilities, not that the paper has to cover _all_ the mentioned points.

We do back up the work with theoretical background and the numerical examples are not even in the presentation from JuliaCon, so these are also new.

I can try to find time to do a literature review and mention what other packages exist that define Lie groups in some sense. I am not aware that this is the case in the Julia package ecosystem; that is why we did not do that until now.
That would take quite some time on my side, since I struggle currently a bit keeping up with all things taking much much longer as I anticipated recently. But that is of course not your fault at all

> [...] perhaps an extended abstract would be a more suitable format?

An extended abstract on the other hand has a 2 page page-limit. Before I would through away the work we put in the current state here and cut that down to 2 pages, I would prefer to not publish at all.

The main idea of the paper was to have a publication to cite this early on for the package to make it a bit more visible and that for a proceedings paper the work load is maybe not that large. We are slowly leaving the work load thing, but that is maybe just life, where nothing works out as one originally has in mind anyways.
