### 2026-01-10 JuliaManifolds/LieGroups.jl#88 @Nimrais
> The main idea of the paper was to have a publication to cite this early on for the package to make it a bit more visible and that for a proceedings paper the work load is maybe not that large. We are slowly leaving the work load thing, but that is maybe just life, where nothing works out as one originally has in mind anyways.

Fair enough. I think the community would definitely benefit more from having your paper published without the benchmark than from having it not published at all—at least because it would make it simpler to cite the package. So on this point, I think we can be lenient. If you find any time to simply cite alternative packages, that would be nice; if not, it is also okay with me.
