### 2026-01-10 JuliaManifolds/LieGroups.jl#88 @Nimrais
> > [...] so if we can compare, we probably should, right?
>
> My personal view point is: I personally can not. But Mateusz might give it a try. I still feel that comparing to exactly one package, that by now was not even on your radar is more like a bit of an ad-hoc comparison. What about C libraries? Fortran? Rust? Do they exist? Can we compare? We for example did a very thorough literature/software review and benchmark comparison in the paper about Manifolds.jl – but that was both a much much longer undertaking as well as a full Journal paper. To me personally this feels out of scope for a package that is as new as LieGroups.jl
>
> > Yeah to be honest, if you can precise your citation in the paper to a specific chapter of the book that would be much less intimidating already.
>
> Currently the paper has 4 placed where we cite Hilgert-Neeb.
>
> * In the introduction at
>
>
>         [LieGroups.jl/paper/paper.tex](https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L59-L60)
>
>
>           Lines 59 to 60
>         in
>         [e10de4c](/JuliaManifolds/LieGroups.jl/commit/e10de4cae7b7fc6a71dfd9718785611b751e4491)
>
>
>
>
>
>
>
>              They have applications in physics, robotics, stochastic processes, information geometry, and many other areas see~\cite{Chirikjian:2009,Chirikjian:2012},
>
>
>
>
>              but are also interesting from their mathematical viewpoint~\cite{HilgertNeeb:2012}
>
>
>
>
>    where besides that book we cite the two books by Chirikjian also without more detail (in sum 800+ pages) since this is meant as an overview
> * In the very first lines of the math section at
>
>
>         [LieGroups.jl/paper/paper.tex](https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L88)
>
>
>            Line 88
>         in
>         [e10de4c](/JuliaManifolds/LieGroups.jl/commit/e10de4cae7b7fc6a71dfd9718785611b751e4491)
>
>
>
>
>
>
>
>              The following notation and definitions follow the text book~\cite{HilgertNeeb:2012},
>
>
>
>
>    together with one reference for manifolds – that is both meant as the generic reference.
> * for the concrete cases of the Lie group definition (
>
>
>         [LieGroups.jl/paper/paper.tex](https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L147)
>
>
>            Line 147
>         in
>         [e10de4c](/JuliaManifolds/LieGroups.jl/commit/e10de4cae7b7fc6a71dfd9718785611b751e4491)
>
>
>
>
>
>
>
>              The (Lie group) exponential (function) \(\exp\colon \mathfrak{g} \to \mathcal{G}\) maps elements from the Lie algebra to the Lie group, and is formally defined~\cite[Def.~9.2.2]{HilgertNeeb:2012} by evaluating the unique curve \(\gamma\colon \mathbb{R} \to \mathcal{G}\) that fulfils the differential equation
>
>
>
>
>    with exact reference to the definition and in the overview for Lie algebra referring to their chapter 5
>
>
>         [LieGroups.jl/paper/paper.tex](https://github.com/JuliaManifolds/LieGroups.jl/blob/e10de4cae7b7fc6a71dfd9718785611b751e4491/paper/paper.tex#L311)
>
>
>            Line 311
>         in
>         [e10de4c](/JuliaManifolds/LieGroups.jl/commit/e10de4cae7b7fc6a71dfd9718785611b751e4491)
>
>
>
>
>
>
>
>              For more details on the Lie bracket see~\cite[Ch.~5]{HilgertNeeb:2012}.
>
>
>
>
>    which – yes – is indeed 40 pages discussing that.
>
> The first is meant as an overview and Chirikjian is on purpose a far more engineering tailored reference as a “counter-weight” – and then even a longer reference. Also – there is no free lunch. Working with Lie groups and Lie algebras does require a bit of diving into the corresponding theory, like it is with many topics. For notation in the second bullet point, I am not sure how to rephrase that, we use the notation of that book. That is what it is; at least the first half of their index can serve as a good notation overview as well.; References 3 and 4 are as precise as I can get them.

I do not have a problem with the first citation, as it is in the Introduction where I think it is acceptable to be less formal with citations without pinpointing them to specific content. Also, I have never read the books by Chirikjian, so it is hard for me to judge that citations.

Citations 3 and 4 are simply good, and I have no problems with them.

The issue I have—which perhaps was not clear from my review—is only with the second citation. Starting immediately in the mathematical background with a reference to the entire book creates the impression that to understand the paper, one first needs to have a good grasp of all the definitions in the book. If something is unclear due to the short format, the reader might feel they need to consult the entire book.

In my view, I do not see why the second citation cannot simply be improved by saying something like: "For readers new to Lie group theory, we encourage consulting Chapters 2 and 9 for comfortable reading" (or whichever chapters are most relevant). That would already be much better.
