### 2026-01-10 JuliaManifolds/LieGroups.jl#88 @kellertuer
> Fair enough. I think the community would definitely benefit more from having your paper published without the benchmark than from having it not published at all—at least because it would make it simpler to cite the package. So on this point, I think we can be lenient. If you find any time to simply cite alternative packages, that would be nice; if not, it is also okay with me.

Thanks for your understanding. To be clear – when the package is a bit more mature, a benchmark would be great to have; I will try to find time to check softwares and literature then.

Concerning the second citation, I will think about what to do, because for now we also do exactly the same for manifolds and a single book we cite. This is also a very very usual thing to do in the beginning of a background/notation section in math papers to cite 2-3 books where such background material is from.
