### 2026-01-11 00adc47
Rephrase Introduction Section 2 to state chapters of the books as well.
