+++ paper/paper.tex
@@ -85,8 +85,9 @@ Finally, in Section~\ref{sec:Example}, we demonstrate how to get started and use
     \label{fig:liegroups_logo}
 \end{figure}
 
-The following notation and definitions follow the text book~\cite{HilgertNeeb:2012},
-for more details on Riemannian manifolds, see also~\cite{DoCarmo:1992}.
+The following notation and definitions follow the text books~\cite{HilgertNeeb:2012},
+especially Chapters 2 and 9 therein.
+For more details on Riemannian manifolds, see also~\cite{DoCarmo:1992} or~\cite{Boumal:2023}, especially Chapter 8 therein.
 
 \subsection{Lie groups}
 We denote a Lie group by \(\mathcal{G} = (\mathcal{M}, \cdot)\) where \(\mathcal{M}\) is a smooth manifold and \(\cdot\) is the group operation.
