### 2026-01-11 JuliaManifolds/LieGroups.jl#88 @kellertuer
Here is my proposal to rephrase the introduction of section 2:

https://github.com/JuliaManifolds/LieGroups.jl/commit/00adc47df2bf20350fd8891c506279a2d4e6d645

Let me know whether that is roughly what you had in mind.

For the overview on other Lie groups packages I plan to do an unnumbered subsection in the introduction, but I first have to check a bit which packages do exist, for now I am only aware of DiffMan (https://github.com/kenthe/DiffMan), which contains Lie groups but is a bit outdated and inactive these days.
