### 2026-01-11 JuliaManifolds/LieGroups.jl#87 @AFeuerpfeil
Hi Ronny,

Thank you for incorporating the changes. The manuscript reads much more clearly now, and I'm happy to approve it for publication.

Best regards,
Andreas
