### 2026-01-11 #195 @AFeuerpfeil
After the revision in https://github.com/JuliaManifolds/LieGroups.jl/issues/87, I am happy to recommend this manuscript for publication.
