### 2026-01-11 JuliaManifolds/LieGroups.jl#87 @kellertuer
Vielen Dank, Andreas!
Also that you did the review so super fast and were available for all the feedback/questions I had.

The I'll close this issue already.
