### 2026-01-11 9086c5d
Collected notes on potentially related packages.
