### 2026-01-11 JuliaManifolds/LieGroups.jl#88 @kellertuer
Evening update: @Nimrais I would like some feedback before continuing.

I collected all packages I could find in

https://github.com/JuliaManifolds/LieGroups.jl/blob/JuliaCon2025Procedings/paper/temp_software_packages.md

For transparency: For exploration I used both google and an LLM, but then checked all packages manually, looked for documentation, features, available groups and citations myself.

My problem here is:

There only seem 3 packages, that go beyond SO/SE(n), n=2,3; these are

* manif
* Sophus, and
*  DiffMan.

 The last is the only one that has a reasonable interface for Lie groups and a comparable amount of Groups similar to LieGroups.jl (the first two have products and Galilean, but not much more).
DiffMan is also a bit outdated (last update 7 years ago, most code 12+ years old, Matlab based).

Also keep in mind that this is a JuliaCon proceedings – so I am not 100% sure how far to look “beyond” Julia.

What do you want to have done to the paper in an overview / comparison for the paper now? I am very very very unsure what to write and what a story could be (also where to put that in the paper).
I do see an inspiration from DiffMan (I also know Kenth and Hans), but since that is Matlab and not maintained, I refrained from citing them for now.

Can you maybe help me, what you would like to read in a comparison to existing packages?
