### 2026-01-11 63d09fd
Add literature
