### 2026-01-12 JuliaManifolds/LieGroups.jl#88 @Nimrais
> Evening update: [@Nimrais](https://github.com/Nimrais) I would like some feedback before continuing.
>
> I collected all packages I could find in
>
> https://github.com/JuliaManifolds/LieGroups.jl/blob/JuliaCon2025Procedings/paper/temp_software_packages.md
>
> For transparency: For exploration I used both google and an LLM, but then checked all packages manually, looked for documentation, features, available groups and citations myself.
>
> My problem here is:
>
> There only seem 3 packages, that go beyond SO/SE(n), n=2,3; these are
>
> * manif
> * Sophus, and
> * DiffMan.
>
> The last is the only one that has a reasonable interface for Lie groups and a comparable amount of Groups similar to LieGroups.jl (the first two have products and Galilean, but not much more). DiffMan is also a bit outdated (last update 7 years ago, most code 12+ years old, Matlab based).
>
> Also keep in mind that this is a JuliaCon proceedings – so I am not 100% sure how far to look “beyond” Julia.
>
> What do you want to have done to the paper in an overview / comparison for the paper now? I am very very very unsure what to write and what a story could be (also where to put that in the paper). I do see an inspiration from DiffMan (I also know Kenth and Hans), but since that is Matlab and not maintained, I refrained from citing them for now.
>
> Can you maybe help me, what you would like to read in a comparison to existing packages?

Thank you for conducting a detailed search of existing packages implementing Lie groups.

As is evident from your review, LieGroups.jl is the only package that is general, actively maintained, and of high quality at the current moment—at least from the ones we can find, and I think it is intellectually honest to phrase it that way :). The other packages are either limited to specific Lie groups, lack documentation, or appear to be unmaintained.

Given this landscape, I think a few lines providing a clear statement of need for this package would be sufficient—perhaps placed between the Introduction and Mathematical Background sections. Since all other packages are clearly not on the same level of quality or generality, a brief mention acknowledging this would adequately address the comparison requirement without requiring the benchmarking.

I suggest following the JuliaCon Proceedings format for a "Statement of Need" section, which could be minimal and read something like (or maybe slightly more polite way regarading the other packages?):

```latex
\subsection{Statement of Need}

Several software packages for working with Lie groups exist across different programming languages, including Sophus and Lie++ (C++), manif and smooth (C++ with Python bindings), jaxlie (Python/JAX), and DiffMan (MATLAB). However, these packages are typically limited to a small set of specific Lie groups---most commonly $\mathrm{SO}(n)$ and $\mathrm{SE}(n)$ for $n = 2, 3$---and lack the generality to define arbitrary Lie groups or extend the interface to new groups. \texttt{LieGroups.jl} addresses this gap by providing both a flexible interface for defining new Lie groups and a comprehensive library of implementations, fully integrated within the \texttt{JuliaManifolds} ecosystem.
```
This would concisely establish the need without requiring performance benchmarks.
