### 2026-01-12 JuliaManifolds/LieGroups.jl#88 @kellertuer
Thanks. I noticed geomstats has a few Lie groups as well, they are just very well hidden ;) So I will add that as well and think a bit how to reference or cite the packages as well (for reproducibility)
