+++ paper/ref.bib
@@ -74,6 +74,27 @@
   TITLE = {Riemannian Geometry},
 }
 
+@article{EngoeMarthinsenMunthe-Kaas:2001,
+  Title = {DiffMan: An object-oriented MATLAB toolbox for solving differential equations on manifolds},
+  Volume = {39},
+  Doi = {10.1016/s0168-9274(00)00042-8},
+  Number = {3?4},
+  Journal = {Applied Numerical Mathematics},
+  Author = {Engø, Kenth and Marthinsen, Arne and Munthe-Kaas, Hans Z.},
+  Year = {2001},
+  Pages = {323?347}
+}
+
+@article{FornasierGeGoorMahonyWeiss:2023,
+  title = {Equivariant symmetries for inertial navigation systems},
+  volume = {181},
+  doi = {10.1016/j.automatica.2025.112495},
+  journal = {Automatica},
+  author = {Fornasier, Alessandro and Ge, Yixiao and van Goor, Pieter and Mahony, Robert and Weiss, Stephan},
+  year = {2025},
+  pages = {112495}
+}
+
 @book{HilgertNeeb:2012,
     AUTHOR    = {Hilgert, Joachim and Neeb, Karl-Hermann},
     DOI       = {10.1007/978-0-387-84794-8},
@@ -89,6 +110,18 @@
   title = {{L}ie{G}roups.jl},
   year = {2025},
   copyright = {MIT License}
+}
+
+@article{MiolaneGuiguiLeBrigantMatheHouThanwerdasHeyderPeltreKoepZaatitiHajriCabanesGeraldChauchatShewmakeBrooksKainzDonnatHolmesPennec:2020,
+  author  = {Nina Miolane and Nicolas Guigui and Alice Le Brigant and Johan Mathe and Benjamin Hou and Yann Thanwerdas and Stefan Heyder and Olivier Peltre and Niklas Koep and Hadi Zaatiti and Hatem Hajri and Yann Cabanes and Thomas Gerald and Paul Chauchat and Christian Shewmake and Daniel Brooks and Bernhard Kainz and Claire Donnat and Susan Holmes and Xavier Pennec},
+  title   = {Geomstats:  A Python Package for Riemannian Geometry in Machine Learning},
+  journal = {Journal of Machine Learning Research},
+  year    = {2020},
+  volume  = {21},
+  number  = {223},
+  pages   = {1--9},
+  url     = {http://jmlr.org/papers/v21/19-027.html},
+  note = {url: \url{http://jmlr.org/papers/v21/19-027.html}.}
 }
 
 @article{Munthe-Kaas:1998,
@@ -110,4 +143,33 @@
   author = {Iserles, Arieh and Munthe-Kaas, Hans Z. and Nørsett, Syvert P. and Zanna, Antonella},
   year = {2000},
   pages = {215–365}
+}
+
+@software{Strasdat:2024,
+  author       = {Strasdat, Hauke},
+  title        = {Sophus: C++ implementation of Lie Groups using Eigen},
+  year         = {2024},
+  url          = {https://github.com/strasdat/Sophus},
+  note = {avaiable on GitHub: \url{https://github.com/strasdat/Sophus}.}
+}
+
+@inproceedings{YiLeeKlossMartinMartinBohg:2021,
+  title = {Differentiable Factor Graph Optimization for Learning Smoothers},
+  doi = {10.1109/iros51168.2021.9636300},
+  booktitle = {2021 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  author = {Yi, Brent and Lee, Michelle A. and Kloss, Alina and Martin-Martin, Roberto and Bohg, Jeannette},
+  year = {2021},
+  pages = {1339?1345}
+}
+
[…]

[…]
