### 2026-01-12 JuliaManifolds/LieGroups.jl#88 @kellertuer
Based on your proposed approach I now wrote

https://github.com/JuliaManifolds/LieGroups.jl/blob/c7cce6a78c2d37e4ee5d5e4b47256c2953107f22/paper/paper.tex#L77-L91

carefully checked the references again and added some that hopefully always either have a DOI or a link the git repository – besides the JLMR paper I am citing, which also only has a URL.

The paper is now neat 8.5 pages long.
