### 2026-01-12 JuliaManifolds/LieGroups.jl#88 @Nimrais
I checked the current manuscript, and I truly believe that this kind of related work section answers a lot of questions about when this package would be interesting for someone to use. I also went through the new citations—they look correct.

I apologize if the review process felt intrusive at times. In hindsight, I could have been more constructive from the start rather than iterating through multiple rounds of feedback. I appreciate your patience and responsiveness throughout this process.

I consider my points to have been addressed. Thank you for the revisions.
