### 2026-01-12 #195 @Nimrais
After the revision in https://github.com/JuliaManifolds/LieGroups.jl/issues/88, I am happy to recommend this manuscript for publication.
