### 2026-01-12 #195 @kellertuer
Super nice! Thanks for the careful review and our discussions on how to improve the paper. I think the new package references were a good idea to add.

Just to make sure the bot is happy, lets also do
