### 2026-01-12 #195 @luraess
Thanks all for your swift actions and constructive comments that significantly improved the manuscript. @kellertuer can you confirm the latest version of the software and share the corresponding (latest) Zenodo link to the archive?

Once I have those, I will generate the final proofs and if we're good to go then it could be wrapped up soon.
