+++ paper/ref.bib
@@ -78,11 +78,11 @@
   Title = {DiffMan: An object-oriented MATLAB toolbox for solving differential equations on manifolds},
   Volume = {39},
   Doi = {10.1016/s0168-9274(00)00042-8},
-  Number = {3?4},
+  Number = {3--4},
   Journal = {Applied Numerical Mathematics},
   Author = {Engø, Kenth and Marthinsen, Arne and Munthe-Kaas, Hans Z.},
   Year = {2001},
-  Pages = {323?347}
+  Pages = {323--347}
 }
 
 @article{FornasierGeGoorMahonyWeiss:2023,
