### 2026-01-13 #195 @kellertuer
That was really all super fast. Thanks also for following up so fast here.

I did check also the paper, since that does cite a zenodo as well. The paper refers to the generic one https://doi.org/10.5281/zenodo.15343362
The current version (0.1.9) mentioned and used in the paper has the specific DOI https://doi.org/10.5281/zenodo.17737448

While working on that, I noticed, there were two `?` in stead of `–` in one reference, so compared to the last rendering above, the final commit is https://github.com/JuliaManifolds/LieGroups.jl/commit/12eb4764606d3d81d1b2ca16cf1063b2a0c6bd6e
but we also plan to always keep the separate paper branch https://github.com/JuliaManifolds/LieGroups.jl/tree/JuliaCon2025Procedings available as well.
Since there was a final commit, should we re-trigger the pdf generation here?
