### 2026-01-13 #195 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
