### 2026-01-13 #195 @luraess
So here we are. To me all looks mostly good. I spotted two minor things one may improve in the final draft, namely:
- Code 1, Line 7: the text goes slightly out of the code box
- Hyperlink colour for references in text are light green which makes it somewhat difficult to read.

If you could maybe address these 2 remaining items and cross-check the final proof is OK for you as well, then we should be good to go.
