### 2026-01-13 3c9b71d
Change hyperref colors to use Paul tols muted scheme, similar to our graphics.
