+++ paper/paper.tex
@@ -27,7 +27,12 @@
 \definecolor{TolMutedWine}{HTML}{882255}
 \definecolor{TolMutedPurple}{HTML}{AA4499}
 %
-\hypersetup{colorlinks=true}
+\hypersetup{%
+    colorlinks=true,%
+    urlcolor=TolMutedRose,%
+    linkcolor=TolMutedBlue,%
+    citecolor=TolMutedGreen%
+}
 %
 \begin{document}
 
