+++ paper/paper.tex
@@ -571,8 +571,8 @@ using CairoMakie
 """
     deCasteljauLieGroup(G, cp, t)
 
-Evaluate the Beziér curve at parameter `t` given
-a vector of control points `cp` on the Lie group `G`.
+Evaluate the Beziér curve at parameter `t` given a
+vector of control points `cp` on the Lie group `G`.
 """
 function deCasteljauLieGroup(
         G::AbstractLieGroup, cp::Vector{P}, t::Real
