### 2026-01-13 #195 @kellertuer
Thanks for spotting those two.

The code line was easy to fix by putting the line break a bit later (the “a” moved to the line 6)

For the colors, we did not change them compared to the given template, but upon your request, I changed the hyper ref colors to the ones proposed by Paul Tol (https://sronpersonalpages.nl/~pault/), where we use the muted colours anyways already in our figures:

that keeps the general idea (red/pink – green – blue) but in dimmed colors for the links.
