### 2026-01-13 #195 @luraess
Excellent - thanks a lot. Can you update the Zenodo thingy to include latest. Then we can finalise and publish - that was fast!
