### 2026-01-13 #195 @kellertuer
The current reference [1] in the paper is the latest DOI (so not the specific 0.1.9 one). Is that what you are referring to? But that is already on latest.

And yes this was fast, from all of you @AFeuerpfeil, @Nimrais and you @luraess – all of you were super fast, helpful and constructive!

_edit:_ or did I just not follow what you mean by “thingy”?
